added timer to teleop_note for output

+++ src/ros2_ws/src/motion_control/src/SerialCommunicator.cpp
@@ -420,7 +420,7 @@ void SerialCommunicator::set_front_speed(int16_t sp)
     // new_speed is in units of RPM (aka - 2pi rad/min)
     unsigned char new_speed[2] {0x00};
     sp *= MC_FRONT_WHEELS_INSTALL_ORIENTATION; 
-    sp *= (120.0 / MC_FRONT_WHEEL_DIAMETER_CM);
+    sp *= int16_t(120.0 / MC_FRONT_WHEEL_DIAMETER_CM);
     convert_int16_to_uchar_(sp, new_speed);
     this->front_wheels_command_[4] = new_speed[0];  // LSB first
     this->front_wheels_command_[5] = new_speed[1];  // MSB second
@@ -433,7 +433,7 @@ void SerialCommunicator::set_rear_speed(int16_t sp)
     // New_speed is in units of RPM (aka - 2pi rad/min)
     unsigned char new_speed[2] {0x00};
     sp *= MC_REAR_WHEELS_INSTALL_ORIENTATION;
-    sp += (120.0 / MC_REAR_WHEEL_DIAMETER_CM);
+    sp += int16_t(120.0 / MC_REAR_WHEEL_DIAMETER_CM);
     convert_int16_to_uchar_(sp, new_speed);
     this->rear_wheels_command_[4] = new_speed[0];  // LSB first
     this->rear_wheels_command_[5] = new_speed[1];  // MSB second
